+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+#  v1.0.1 (2017-01-24)
+
+### Features
+
+* **More control over options**
+
+### Bug Fixes
+
+* URLs with HTML errors are parsed but not minified. 
+
+### BREAKING CHANGES
+
+* none
+
 #  v1.0.0 (2017-01-19)
 
 ### Features